requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
requests==2.32.5
pydantic==2.11.7
python-dotenv==1.1.1